+++ TidalPy/utilities/dimensions/nondimensional.pyx
@@ -12,7 +12,7 @@ Martens16 : H. Martens, PhD Thesis (CalTech), 2016, DOI: 10.7907/Z9N29TX7
 
 from libc.math cimport sqrt
 
-from TidalPy.constants cimport d_G, d_PI_DBL, d_NAN_DBL
+from TidalPy.constants cimport d_G, d_PI, d_NAN_DBL
 
 
 cdef class NonDimensionalScalesClass:
@@ -66,7 +66,7 @@ cdef void cf_build_nondimensional_scales(
         double bulk_density
         ) noexcept nogil:
 
-    non_dim_scales_ptr.second2_conversion = 1. / (d_PI_DBL * d_G * bulk_density)
+    non_dim_scales_ptr.second2_conversion = 1. / (d_PI * d_G * bulk_density)
     non_dim_scales_ptr.second_conversion  = sqrt(non_dim_scales_ptr.second2_conversion)
     non_dim_scales_ptr.length_conversion  = mean_radius
     non_dim_scales_ptr.length3_conversion = mean_radius * mean_radius * mean_radius